Using Closed Type Families now.
As we depend on opaleye-trans, and that has a base >= 4.8 constraint,
and as this package was never truly compatible with older GHCs, we give
up on a lot of compatibility shite in this commit.

We've also changed the type of the writer columns given by `optional`
from `Maybe` to `Option`. `Option` is just a newtype for `Maybe` that we
define ourselves. This allows us to distinguish between a nullable column (for
which we use `Maybe`), and an optional writer column, for which we use
`Option`.