Major overhaul of Opaleye.StreamBeneathTheHills.Table
First of all, `table` and and `tableWithSchema` now take a "mangler" function, which makes them much more useful. The function computes the name of the database column corresponding to the record field. This makes them much more practically useful.

We also define a synonym for `Opaleye.Table`, which simply takes only a single argument (which is the Haskell-valued writer columns), and then type families are used to calculate the reader columns and postgres-ify the Haskell values. Alongside this is the `TableArgs` type class to help type inference.