I had butchered the Select profunctor
We want the projection function to be a separate function, not just lmap'd onto the QueryRunner. This is because when we go to actually run the query, we want to rmap it onto the Query itself, not the QueryRunner, so that we don't end up pulling unnecessary fields from the database.